+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Added check for Duplicate LevelLoader objects
   - Added check for duplicate loaded levels
   - Added check for correct level load flows
+- Added `SPACE.cs` enum to allow selection between World & Local spaces
+- Added `WaitForRotationAnimation.cs` to adjust tranform rotations easily
 
 ### Changed
 - Updated `LevelLoader.cs` to improve its usability
@@ -19,6 +21,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - Added editor only debugging to load specific levels
   - Adjusted `LoadFirstLevel()` to load the debug index when in editor, and the user has enabled debugging
   - Added `LoadLevelAtIndex()` to simplify external access to loading levels
+- Added functionality to `WaitForAnimationBase.cs` to be called from Unity Events
+  - Added `WaitForAnimationBase.Animate()` to be easily called from Unity Events
+  - Added `defaultAnimationTime` to allow `Animate()` to use when calling
+  - Added `m_lastAnimationDirection` to allow `Animate()` to ping-pong between positions when calling `Animate()`
+- Added Transform Space selection for WaitForAnimations
+  - Added SPACE field into `WaitForAnimationBase.AnimationData` class
+  - Added SPACE switch into `WaitForMoveAnimations.SetValue()`
 
 ### Fixed
 - a